requirements.txt
Flask==2.0.2
mysql-connector-python==8.0.27
pyTelegramBotAPI==4.2.0
requests==2.27.1

